Edit view livewire crm department-list

+++ resources/views/livewire/department/department-lists.blade.php
@@ -39,10 +39,14 @@
                 </div>
                 <div class="col-3 d-flex justify-content-center">
                     <div class="btn-group w-100" role="group">
-                        {{-- <button wire:click="$dispatch('add-department')" type="button" class="btn btn-primary"
-                            data-toggle="modal" data-target="#modal-add-department">
-                            <i class="fas fa-plus"></i> Department
-                        </button> --}}
+
+                        @can('department.create')
+                            <button wire:click="$dispatch('add-department')" type="button" class="btn btn-primary"
+                                data-toggle="modal" data-target="#modal-add-department">
+                                <i class="fas fa-plus"></i> Department
+                            </button>
+                        @endcan
+
                         <button wire:click="$dispatch('refresh-department')" type="button" class="btn btn-success">
                             <i class="fas fa-sync-alt"></i> Refresh
                         </button>
@@ -68,13 +72,18 @@
         <!-- /.card-header -->
         <div class="card-body">
             <div class="table-responsive">
-                <table class="table table-hover">
-                    <thead>
+                <table class="table table-sm table-hover table-bordered">
+                    <thead class="table-info">
                         <th scope="col">#</th>
                         <th scope="col">Created</th>
                         <th scope="col">Updated</th>
+                        <th scope="col">Id</th>
                         <th scope="col">Department Name</th>
-                        <th scope="col">Action</th>
+                        @can('department.edit')
+                            <th scope="col" colspan="2">Action</th>
+                        @elsecan('department.delete')
+                            <th scope="col" colspan="2">Action</th>
+                        @endcan
                     </thead>
 
                     <tbody>
@@ -117,19 +126,30 @@
                                     </div>
 
                                 </td>
+                                <td>{{ $item->id }}</td>
                                 <td>{{ $item->name }}</td>
-                                <td>
-                                    {{-- <button wire:click.prevent="$dispatch('edit-department',{id:{{ $item->id }}})"
-                                        type="button" class="btn btn-secondary btn-sm disabled" data-toggle="modal"
-                                        data-target="#modal-edit-department">
-                                        <i class="fas fa-edit"></i>
-                                    </button>
-                                    <button
-                                        wire:click.prevent="deleteDepartment({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
-                                        class="btn btn-sm btn-danger disabled">
-                                        <i class="fas fa-trash"></i>
-                                    </button> --}}
-                                </td>
+
+                                @can('department.edit')
+                                    <td style="width: 45px" class="p-1 text-center">
+                                        <button wire:click.prevent="$dispatch('edit-department',{id:{{ $item->id }}})"
+                                            type="button" class="btn btn-secondary btn-sm" data-toggle="modal"
+                                            data-target="#modal-edit-department">
+                                            <i class="fas fa-edit"></i>
+                                        </button>
+
+                                    </td>
+                                @endcan
+
+                                @can('department.delete')
+                                    <td style="width: 45px" class="p-1 text-center">
+                                        <button
+                                            wire:click.prevent="deleteDepartment({{ $item->id }},{{ "'" . str_replace("'", '', $item->name) . "'" }})"
+                                            class="btn btn-sm btn-danger">
+                                            <i class="fas fa-trash"></i>
+                                        </button>
+                                    </td>
+                                @endcan
+
                             </tr>
                         @endforeach
                         @livewire('department.department-edit')
